src/main/java/de/workshops/bookdemo/book/BookRepository.java
package de.workshops.bookdemo.book;

import java.io.IOException;
import java.sql.ResultSet;
import java.sql.SQLException;
import java.util.List;

import com.fasterxml.jackson.core.type.TypeReference;
import com.fasterxml.jackson.databind.ObjectMapper;

import org.springframework.core.io.ResourceLoader;
import org.springframework.jdbc.core.JdbcTemplate;
import org.springframework.jdbc.core.RowMapper;
import org.springframework.stereotype.Repository;

import lombok.RequiredArgsConstructor;

@Repository
@RequiredArgsConstructor
public class BookRepository {

    private final ObjectMapper mapper;
    
    private final ResourceLoader resourceLoader;
    private final JdbcTemplate jdbcTemplate;

    private List<Book> books;

    


    //@PostConstruct
    public void init() throws IOException {
        final var resource = resourceLoader.getResource("classpath:books.json");
        this.books = mapper.readValue(resource.getInputStream(), new TypeReference<>() {});
    }

    public List<Book> findAllBooks() {
        return jdbcTemplate.query("SELECT * FROM book", new RowMapper<Book>() {

            @Override
            public Book mapRow(ResultSet rs, int rowNum) throws SQLException {
                Book book = new Book();
                book.setTitle(rs.getString("title"));

                return book;
            }

        });
    }

}
